feat(cb2-8459): test hgv
