+++ json-definitions/v3/tech-record/get/hgv/testable/index.json
@@ -54,7 +54,7 @@
         "techRecord_adrDetails_compatibilityGroupJ": {
             "type": [
                 "boolean",
-                null
+                "null"
             ]
         },
         "techRecord_adrDetails_additionalExaminerNotes": {
@@ -102,7 +102,7 @@
         "techRecord_adrDetails_memosApply": {
             "type": [
                 "array",
-                null
+                "null"
             ],
             "items": {
                 "type": [
@@ -113,7 +113,7 @@
         "techRecord_adrDetails_documents": {
             "type": [
                 "array",
-                null
+                "null"
             ],
             "items": {
                 "type": [
@@ -323,7 +323,7 @@
         },
         "techRecord_alterationMarker": {
             "type": [
-                "string",
+                "boolean",
                 "null"
             ]
         },
@@ -561,7 +561,7 @@
         },
         "techRecord_euroStandard": {
             "type": [
-                "boolean",
+                "string",
                 "null"
             ]
         },
@@ -801,7 +801,7 @@
             "type": "string"
         },
         "techRecord_recordCompleteness": {
-            "const": "skeleton"
+            "const": "testable"
         },
         "techRecord_regnDate": {
             "anyOf": [
@@ -820,7 +820,7 @@
         },
         "techRecord_roadFriendly": {
             "type": [
-                "string",
+                "boolean",
                 "null"
             ]
         },
@@ -940,7 +940,7 @@
             ]
         },
         "techRecord_vehicleType": {
-            "const": "trl"
+            "const": "hgv"
         },
         "primaryVrm": {
             "type": "string"
