+++ json-schemas/v3/tech-record/get/hgv/skeleton/index.json
@@ -50,7 +50,7 @@
 		"techRecord_adrDetails_compatibilityGroupJ": {
 			"type": [
 				"boolean",
-				null
+				"null"
 			]
 		},
 		"techRecord_adrDetails_additionalExaminerNotes": {
@@ -98,7 +98,7 @@
 		"techRecord_adrDetails_memosApply": {
 			"type": [
 				"array",
-				null
+				"null"
 			],
 			"items": {
 				"type": [
@@ -109,7 +109,7 @@
 		"techRecord_adrDetails_documents": {
 			"type": [
 				"array",
-				null
+				"null"
 			],
 			"items": {
 				"type": [
@@ -329,7 +329,7 @@
 		},
 		"techRecord_alterationMarker": {
 			"type": [
-				"string",
+				"boolean",
 				"null"
 			]
 		},
@@ -390,7 +390,10 @@
 			"maxLength": 255
 		},
 		"techRecord_applicationId": {
-			"type": "string"
+			"type": [
+				"string",
+				"null"
+			]
 		},
 		"axles": {
 			"type": "array",
@@ -572,7 +575,7 @@
 		},
 		"techRecord_euroStandard": {
 			"type": [
-				"boolean",
+				"string",
 				"null"
 			]
 		},
@@ -934,7 +937,7 @@
 		},
 		"techRecord_roadFriendly": {
 			"type": [
-				"string",
+				"boolean",
 				"null"
 			]
 		},
@@ -1109,7 +1112,7 @@
 			]
 		},
 		"techRecord_vehicleType": {
-			"const": "trl"
+			"const": "hgv"
 		},
 		"primaryVrm": {
 			"type": "string"
